refactor(nuxt): move nuxt island into its own event handler
