+++ packages/nuxt/src/core/runtime/nitro/utils/renderer/inline-styles.ts
@@ -1,0 +1,15 @@
+import type { Style } from '@unhead/vue/types'
+import { getSSRStyles } from './build-files'
+
+export async function renderInlineStyles (usedModules: Set<string> | string[]): Promise<Style[]> {
+  const styleMap = await getSSRStyles()
+  const inlinedStyles = new Set<string>()
+  for (const mod of usedModules) {
+    if (mod in styleMap && styleMap[mod]) {
+      for (const style of await styleMap[mod]()) {
+        inlinedStyles.add(style)
+      }
+    }
+  }
+  return Array.from(inlinedStyles).map(style => ({ innerHTML: style }))
+}
