+++ packages/nuxt/src/core/runtime/nitro/utils/renderer/payload.ts
